+++ src/Application.php
@@ -4,6 +4,7 @@ namespace Laravel\Lumen;
 
 use Closure;
 use Exception;
+use RuntimeException;
 use Illuminate\Support\Str;
 use Illuminate\Container\Container;
 use Illuminate\Support\Facades\Facade;
@@ -135,7 +136,17 @@ class Application extends Container implements ApplicationContract
      */
     public function version()
     {
-        return 'Lumen (5.3.0-dev) (Laravel Components 5.3.*)';
+        return 'Lumen (5.3.0) (Laravel Components 5.3.*)';
+    }
+
+    /**
+     * Determine if the application is currently down for maintenance.
+     *
+     * @return bool
+     */
+    public function isDownForMaintenance()
+    {
+        return file_exists($this->storagePath('framework/down'));
     }
 
     /**
@@ -162,16 +173,6 @@ class Application extends Container implements ApplicationContract
         }
 
         return $env;
-    }
-
-    /**
-     * Determine if the application is currently down for maintenance.
-     *
-     * @return bool
-     */
-    public function isDownForMaintenance()
-    {
-        return file_exists($this->storagePath('framework/down'));
     }
 
     /**
@@ -225,10 +226,12 @@ class Application extends Container implements ApplicationContract
 
         $this->loadedProviders[$providerName] = true;
 
-        $provider->register();
+        if (method_exists($provider, 'register')) {
+            $provider->register();
+        }
 
         if (method_exists($provider, 'boot')) {
-            $this->call([$provider, 'boot']);
+            return $this->call([$provider, 'boot']);
         }
     }
 
@@ -415,11 +418,18 @@ class Application extends Container implements ApplicationContract
     /**
      * Register the facades for the application.
      *
-     * @return $this
+     * @param  bool  $aliases
+     * @param  array $custom
+     *
+     * @return void
      */
-    public function withFacades()
+    public function withFacades($aliases = true, $custom = [])
     {
         Facade::setFacadeApplication($this);
+
+        if ($aliases) {
+            $this->withAliases($custom);
+        }
 
         return $this;
     }
@@ -428,28 +438,48 @@ class Application extends Container implements ApplicationContract
      * Register the facade aliases for the application.
      *
      * @return $this
+     *
+     * @deprecated v3.3.0
      */
     public function withFacadeAliases()
     {
+        $this->withAliases();
+    }
+
+    /**
+     * Register the aliases for the application.
+     *
+     * @param  array  $custom
+     *
+     * @return void
+     */
+    public function withAliases($custom = [])
+    {
+        $defaults = [
+            'Illuminate\Support\Facades\Auth'      => 'Auth',
+            'Illuminate\Support\Facades\Cache'     => 'Cache',
+            'Illuminate\Support\Facades\DB'        => 'DB',
+            'Illuminate\Support\Facades\Crypt'     => 'Crypt',
+            'Illuminate\Support\Facades\Event'     => 'Event',
+            'Illuminate\Support\Facades\Gate'      => 'Gate',
+            'Illuminate\Support\Facades\Hash'      => 'Hash',
+            'Illuminate\Support\Facades\Log'       => 'Log',
+            'Illuminate\Support\Facades\Queue'     => 'Queue',
+            'Illuminate\Support\Facades\Schema'    => 'Schema',
+            'Illuminate\Support\Facades\Session'   => 'Session',
+            'Illuminate\Support\Facades\Storage'   => 'Storage',
+            'Illuminate\Support\Facades\URL'       => 'URL',
+            'Illuminate\Support\Facades\Validator' => 'Validator',
+        ];
+
         if (! static::$aliasesRegistered) {
             static::$aliasesRegistered = true;
 
-            class_alias('Illuminate\Support\Facades\App', 'App');
-            class_alias('Illuminate\Support\Facades\Auth', 'Auth');
-            class_alias('Illuminate\Support\Facades\DB', 'DB');
-            class_alias('Illuminate\Support\Facades\Cache', 'Cache');
-            class_alias('Illuminate\Support\Facades\Crypt', 'Crypt');
-            class_alias('Illuminate\Support\Facades\Event', 'Event');
-            class_alias('Illuminate\Support\Facades\Gate', 'Gate');
-            class_alias('Illuminate\Support\Facades\Hash', 'Hash');
-            class_alias('Illuminate\Support\Facades\Log', 'Log');
-            class_alias('Illuminate\Support\Facades\Mail', 'Mail');
-            class_alias('Illuminate\Support\Facades\Queue', 'Queue');
-            class_alias('Illuminate\Support\Facades\Request', 'Request');
-            class_alias('Illuminate\Support\Facades\Session', 'Session');
-            class_alias('Illuminate\Support\Facades\Storage', 'Storage');
-            class_alias('Illuminate\Support\Facades\URL', 'URL');
-            class_alias('Illuminate\Support\Facades\Validator', 'Validator');
+            $merged = array_merge($defaults, $custom);
+
+            foreach ($merged as $original => $alias) {
+                class_alias($original, $alias);
+            }
         }
 
         return $this;
@@ -560,12 +590,40 @@ class Application extends Container implements ApplicationContract
     /**
      * Prepare the application to execute a console command.
      *
+     * @param  bool  $aliases
+     *
      * @return void
      */
-    public function prepareForConsoleCommand()
+    public function prepareForConsoleCommand($aliases = true)
     {
-        $this->withFacades();
+        $this->withFacades($aliases);
 
         $this->configure('database');
+    }
+
+    /**
+     * Get the application namespace.
+     *
+     * @return string
+     *
+     * @throws \RuntimeException
+     */
+    public function getNamespace()
+    {
+        if (! is_null($this->namespace)) {
+            return $this->namespace;
+        }
+
+        $composer = json_decode(file_get_contents($this->basePath('composer.json')), true);
+
+        foreach ((array) data_get($composer, 'autoload.psr-4') as $namespace => $path) {
+            foreach ((array) $path as $pathChoice) {
+                if (realpath($this->path()) == realpath($this->basePath().'/'.$pathChoice)) {
+                    return $this->namespace = $namespace;
+                }
+            }
+        }
+
+        throw new RuntimeException('Unable to detect application namespace.');
     }
 }