+++ src/Concerns/RoutesRequests.php
@@ -6,6 +6,7 @@ use Closure;
 use Exception;
 use Throwable;
 use FastRoute\Dispatcher;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Laravel\Lumen\Routing\Pipeline;
@@ -191,6 +192,7 @@ trait RoutesRequests
      * @param  array|string  $method
      * @param  string  $uri
      * @param  mixed  $action
+     *
      * @return void
      */
     public function addRoute($method, $uri, $action)
@@ -333,7 +335,13 @@ trait RoutesRequests
      */
     public function handle(SymfonyRequest $request, $type = HttpKernelInterface::MASTER_REQUEST, $catch = true)
     {
-        return $this->dispatch($request);
+        $response = $this->dispatch($request);
+
+        if (count($this->middleware) > 0) {
+            $this->callTerminableMiddleware($response);
+        }
+
+        return $response;
     }
 
     /**
@@ -367,6 +375,10 @@ trait RoutesRequests
      */
     protected function callTerminableMiddleware($response)
     {
+        if ($this->shouldSkipMiddleware()) {
+            return;
+        }
+
         $response = $this->prepareResponse($response);
 
         foreach ($this->middleware as $middleware) {
@@ -374,7 +386,7 @@ trait RoutesRequests
                 continue;
             }
 
-            $instance = $this->make($middleware);
+            $instance = $this->make(explode(':', $middleware)[0]);
 
             if (method_exists($instance, 'terminate')) {
                 $instance->terminate($this->make('request'), $response);
@@ -545,7 +557,13 @@ trait RoutesRequests
      */
     protected function callControllerAction($routeInfo)
     {
-        list($controller, $method) = explode('@', $routeInfo[1]['uses']);
+        $uses = $routeInfo[1]['uses'];
+
+        if (is_string($uses) && ! Str::contains($uses, '@')) {
+            $uses .= '@__invoke';
+        }
+
+        list($controller, $method) = explode('@', $uses);
 
         if (! method_exists($instance = $this->make($controller), $method)) {
             throw new NotFoundHttpException();
@@ -650,10 +668,7 @@ trait RoutesRequests
      */
     protected function sendThroughPipeline(array $middleware, Closure $then)
     {
-        $shouldSkipMiddleware = $this->bound('middleware.disable') &&
-                                        $this->make('middleware.disable') === true;
-
-        if (count($middleware) > 0 && ! $shouldSkipMiddleware) {
+        if (count($middleware) > 0 && ! $this->shouldSkipMiddleware()) {
             return (new Pipeline($this))
                 ->send($this->make('request'))
                 ->through($middleware)
@@ -717,5 +732,15 @@ trait RoutesRequests
     public function getRoutes()
     {
         return $this->routes;
+    }
+
+    /**
+     * Determines whether middleware should be skipped during request.
+     *
+     * @return bool
+     */
+    protected function shouldSkipMiddleware()
+    {
+        return $this->bound('middleware.disable') && $this->make('middleware.disable') === true;
     }
 }