+++ docs/changes.md
@@ -5,6 +5,21 @@ title: Lumen Change Log
 
 ## Version 3.2 {#v3-2}
 
+### v3.2.8 {#v3-2-8}
+
+* Add `Laravel\Lumen\Application::getNamespace()`. ([@mlantz](https://github.com/mlantz))
+* Fixes unit testing Jobs using mock object always fails. ([@christian452](https://github.com/christian452))
+* Add support for handling transformer and serializer using `Laravel\Lumen\Routing\Helpers` trait.
+
+### v3.2.7 {#v3-2-7}
+
+* Add support for installation via `lumen.json`.
+
+### v3.2.6 {#v3-2-6}
+
+* Improved the exception handler. ([@GrahamCampbell](https://github.com/GrahamCampbell))
+* Add `Laravel\Lumen\Application::runningUnitTests()`.
+
 ### v3.2.5 {#v3-2-5}
 
 * Allow middleware to replace route info. ([@sw-double](https://github.com/sw-double))
@@ -49,6 +64,10 @@ title: Lumen Change Log
 * Add `symfony/polyfill-php56` to add support `hash_equals()` on < PHP 5.6.
 
 ## Version 3.1 {#v3-1}
+
+### v3.1.9 {#v3-1-9}
+
+* Improved the exception handler. ([@GrahamCampbell](https://github.com/GrahamCampbell))
 
 ### v3.1.8 {#v3-1-8}